Comments and refactor
